Save invalid quizes as drafts
If an article has a draft, load it in the editor.
If an article has a public quiz, load it in the front end.
If a valid quiz is saved, drop the draft and save the quiz as a public quiz.
If an invalid quiz is saved, keep the public version and save the quiz as a draft.